-moved testing complex mechanisms into a sep file
