+++ tests/test_simple_market_mechanisms.py
@@ -1,0 +1,95 @@
+from datetime import datetime, timedelta
+
+from dateutil import rrule as rr
+from dateutil.relativedelta import relativedelta as rd
+
+from assume.common.market_objects import MarketConfig, MarketProduct
+from assume.common.utils import get_available_products
+from assume.markets import MarketRole, clearing_mechanisms
+
+from .utils import create_orderbook, extend_orderbook
+
+simple_dayahead_auction_config = MarketConfig(
+    "simple_dayahead_auction",
+    market_products=[MarketProduct(rd(hours=+1), 1, rd(hours=1))],
+    additional_fields=["node_id"],
+    opening_hours=rr.rrule(
+        rr.HOURLY,
+        dtstart=datetime(2005, 6, 1),
+        cache=True,
+    ),
+    opening_duration=timedelta(hours=1),
+    volume_unit="MW",
+    volume_tick=0.1,
+    price_unit="€/MW",
+    market_mechanism="pay_as_clear",
+)
+
+
+def test_market():
+    mr = MarketRole(simple_dayahead_auction_config)
+    next_opening = simple_dayahead_auction_config.opening_hours.after(datetime.now())
+    products = get_available_products(
+        simple_dayahead_auction_config.market_products, next_opening
+    )
+    assert len(products) == 1
+
+    print(products)
+
+    """
+    Create Orderbook with constant order volumes and prices:
+        - dem1: volume = -1000, price = 3000
+        - gen1: volume = 1000, price = 100
+        - gen2: volume = 900, price = 50
+    """
+    orderbook = extend_orderbook(products, -1000, 3000)
+    orderbook = extend_orderbook(products, 1000, 100, orderbook)
+    orderbook = extend_orderbook(products, 900, 50, orderbook)
+
+    simple_dayahead_auction_config.market_mechanism = clearing_mechanisms[
+        simple_dayahead_auction_config.market_mechanism
+    ]
+    mr.all_orders = orderbook
+    accepted, rejected, meta = simple_dayahead_auction_config.market_mechanism(
+        mr, products
+    )
+    assert meta[0]["demand_volume"] > 0
+    assert meta[0]["price"] > 0
+    import pandas as pd
+
+    print(pd.DataFrame(mr.all_orders))
+    print(pd.DataFrame(accepted))
+    print(meta)
+
+
+def test_simple_market_mechanism():
+    import copy
+
+    for name, mechanism in clearing_mechanisms.items():
+        if "complex" in name:
+            continue
+
+        print(name)
+        market_config = copy.copy(simple_dayahead_auction_config)
+        market_config.market_mechanism = mechanism
+        mr = MarketRole(market_config)
+        next_opening = market_config.opening_hours.after(datetime.now())
+        products = get_available_products(market_config.market_products, next_opening)
+        assert len(products) == 1
+        order = {
+            "start_time": products[0][0],
+            "end_time": products[0][1],
+            "only_hours": products[0][2],
+        }
+
+        orderbook = create_orderbook(order, node_ids=[0, 1, 2])
+        mr.all_orders = orderbook
+        accepted, rejected, meta = market_config.market_mechanism(mr, products)
+        assert meta[0]["supply_volume"] > 0
+        assert meta[0]["price"] > 0
+        # import pandas as pd
+        # print(pd.DataFrame(mr.all_orders))
+        # print(pd.DataFrame(clearing_result))
+        # print(meta)
+
+    # return mr.all_orders, meta
